Add specs for the stack itself working
Previously, the stack itself did not have tests, and more than
once I've broken it and had to rely on reports from users to fix
it.  This adds basic specs for the stack itself, where it clones
the stack to a subdirectory, runs rake setup, then runs various
rake commands to make sure the stack functions.  This should
significantly reduce the odds of future stack breakage.