Fix trivial crash
But, now the problem is the way minitest and rspec prefer to gather
tests are different.  Next is to dig into the Rakefile or maybe
consider using other rspec features.  It's kind of interesting how the
roda-sequel-stack lets you be optimal in a way by changing why global
features you have (e.g. rack/test only loaded for web tests).