Turn off some rspec cops
Though I am grateful that rspec rubocop taught me about spies, these
three rules are wrongheaded in the context of large test suites
targeting very high coverage.  They are explained one at a time:

RSpec/MultipleExpectations: wrongheaded, once you've set up a
non-trivial fixture, you want to get multiple invariants checked at
once for performance.

RSpec/StubbedMock: wrongheaded because "expect" not only has use to
check invariants, but also as a form of partial doubling *that can
alert you when the stub is no longer needed.* Or, if you made a
mistake in the test and really do expect the message to be recieved
(see previous).

"allow" is not adequate: it will accumulate as a plaque as the data
dependencies change.  People read the test, assuming an obsolete
"allow" does something, yet, it doesn't. "allow" is nevertheless
useful if you need to make a shared fixture.

RSpec/MessageSpies: wrongheaded because it complains about *any* use
of partial doubling.  Okay, sure "consider" the spy, but complain each
and every time about it?  This cop is useful exactly once to educate
the person who turns it off (thanks, by the way).