Add the first label execution and test for it
A label is simply a method, but it is evocative of the same construct
in C or assembly programs: jumping to a code position without
implications for the "stack," which has yet to be fleshed out.

Most linear control flow in Clover will tend to be label jumps, and
only composed programs will use the stack.