Fix a bug the Strand would crash if deleted
When a strand deletes itself, as in the case of a root Strand that
cannot rely on a parent strand to reap it, the leasing defect
detection code delivers a bogus positive.  This adds code to avoid
that.

It would be nice if Sequel model code tracked deletion state (maybe:
there's some tricky interactions with transactions that could result),
but it doesn't seem to, so we expend complications to do so ourselves.