Add Sshable model
This can be used for host servers, or possibly VMs if Clover ever uses
both in some capacity.  It doesn't matter, as long as there's a host.

Included is field encryption and connection caching.

Note that the connection cache has no limit, and is bound to a
Thread. The intention is that Threads made in the Dispatcher should be
short-lived, rather than pooled and reused.

Previous projects of mine have tended to make profuse use of SSH
connections to simplify the programming model, but this was slow.  By
binding the cache to the Thread, the programming model stays simple,
and hopefully the caching code stays simple, and we are not stuck
debugging complex residue across many executions in a long-lived
thread.