Integrate a base64 decode into column encryption
Although Ruby lets you put binary data into the ENV hash (which is how
I was using this), binary data in environment variables is not
something Linux actually supports because the environment page is
NUL-separated.

And it's inconvenient.  So imply a base64 encoded spring in the key
serialization where it is read.