Several Config enhancements
Make some fields optional, because, for example:

* Migration scripts need CLOVER_DATABASE_URL but not
  CLOVER_SESSION_SECRET nor CLOVER_COLUMN_ENCRYPTION_KEY.

* bin/respirate also does not need CLOVER_SESSION_SECRET

The expedient support for ENV clearing that I adapted from
`roda-sequel-stack` was only for mandatory fields, I out-lined the
common features into a block and made it work for `optional` fields as
well.

Finally, centralize base64 decoding as a type of "casting" in Config,
to clean up code that uses Config.