Add apoptosis
To make leases work, the process and threads have to end before the
LEASE_EXPIRATION is reached.

I do this with Ractors, in the past, I've done so with Threads.
Although Threads have not failed me, in principle, they could, if any
C library were to fail to yield to Ruby while sitting on the GVL.