Abstract the virtual machine setup into a class
By doing this, it's easier to recombine steps in the process.  The
most relevant use of this is in development, where it can be useful to
load a REPL and re-execute slices of the VM process.

This is apparent under some testing scenarios, when it's useful to
clear all network namespace and entities via reboot.  In such a case,
there will already be a VM user and various resource files, so running
a subset of the network setup code is useful.