Integrate systemd unit installation into prep and purge
This is the new way to start VMs, mostly because it's amenable to
advanced systemd containment options.

In fact, the old `exec_cloud_hypervisor`/`runvm.rb` method can be
considered (relatively) insecure, especially once we pile on more
systemd options to constrain the process it runs:

https://www.redhat.com/sysadmin/mastering-systemd

...and so, we should probably get rid of runvm.rb and its mechanism,
excepting as a development aid, where a foregrounded serial tty can be
useful.