Release lease in presence of exception
It was highly annoying, and unintentional, that raising an exception
would skip lease-release code.  Thus, in development, you'd `.run` buggy
code and then have to update the lease to nil yourself to try again
right away.

This is not the design intention of leases, which should end the moment
code can certify that it is done working, with or without an exception.