Add Sshable#cmd to run SSH commands
This method returns the stdout of the run command and raises an error
if it has a non-zero exit code.  To halt execution is most often the
best default behavior in the context of Clover, where Progs are likely
to perform re-executions when impeded.

Two RSpec RuboCop rules are disabled:

RSpec/ExpectInHook: uses of expect in a hook is situational, because
it means all tests follow a particular code path rather precisely.  At
such a time that is too burdensome, this can be degraded to `allow`,
which carries the downside that un-used `allow` is sclerotic, prone to
accumulation even if unused and misleading readers.  See also:
d5b49bb05bb3e88cb3d1b5e2742ed31e3e20cb26

RSpec/SubjectStub: I understand the theory of not changing the objects
under test, but in equilibrium I find abiding this prohibition in
every case bloats code or reduces the test coverage.  I prefer test
churn or occasional mistakes to those, so I am disabling the rule.