Handful of table adjustments
Suspending the usual rules about editing migrations (again), I make a
few changes for convenience sake in this early period:

`parent` is renamed to `parent_id`.  This is to better abide Sequel's
association support, which will want to use the "parent" name for its
methods.

`schedule` is given a default.  It cleans up many tests, and probably
code in production as well.  The code that wants to hint at suspending
execution for a while can do so after executing once.

`retval` has been made NULLable, `exitval` has been added: it occurred
to me that I shouldn't conflate Prog exit and stack returns, and that
a NULL state is useful to indicate "hasn't returned" and "hasn't
exited," a special empty value is more prone to error here.