Implement Prog composition and host bootstrapping
This patch allows preparing a blank Hetzner (or other root-ssh-keyed)
instance automatically.  First, create a `Sshable` record and a
`VmHost` record.  Note how they are correlated by sharing an `id`:

    # In development, ssh-agent is used rather than database-stored private keys for now
    sa = Sshable.create(host: '95.216.67.38', private_key: '')
    vh = VmHost.create { _1.id = sa.id }

Then create a "Strand" (compare: Thread, Fiber) running the
`VmHostNexus` prog:

    st = Strand.create(prog: "VmHostNexus", label: "start") {
      _1.id = vh.id
    }

Then, you can run:

    st.run

Until the `VmHostNexus` ends in the `wait` label, bootstrapping the
host along the way.

== How is this organized

Clover uses metaphors from machine and operating system architecture
to organize its asynchronous processing:

* Strands are like Threads or Fibers
* Progs are like program text. They get loaded into strands.
* Parent strands, like unix processes, are responsible for reaping
  child strands
* The stack is like...the stack
* There are return values (when popping stack) and exit codes/values
  (when ending the strand)

Since this patch also introduces composition of Progs to organize
this, I will go over how it works.

First, there's some convention: there is a category of Prog, called
the "Nexus."  In this patch, it is the `VmHostNexus`.  The Nexus prog
is special:

* It runs on a root Strand, with no parent.  Its lifetime is roughly
  as long as the entity record itself, i.e. from creation to deletion.
* A root strand shares an id with an entity record (`VmHost`).  This
  allows you to write `VmHost.first.strand.run`, for example.
* A root strand primarily exists to start descendant progs
* A root strand waits for those progs to complete before moving on

The way those descendant progs are started is via the `bud` method.
The `bud` method accepts an initial stack frame, to be encoded as
JSONB, as a parameter.  Most of the time, initial stack contains the
id of an entity record.

N.B.: As of yet, we have no code that pushes frames onto the stack,
      save for initialization in `bud`.  The future design intention
      is that, in cases where concurrency is not desired, instead of
      creating a child strand we could push and pop frames on the
      stack instead to compose progs.  I do not envision stacks, or
      Strand parent relationships, being deeper than four elements,
      however.

The way that child strands are cleared is via `reap`.

After reaping, `leaf?` is used to determine if there are any
descendant progs left.

Within a prog, progress is recorded via `hop`, which is inspired by
the machine code idea of "jump".

`translocate` is used as a development convenience that should also be
harmless for some time in production: it allows `.run` on a `Strand`
(in practice, a root strand) to execute code in descendant strands.
This is useful when the parent strand is waiting for child strands to
exit and has nothing else to do.  This is what makes the loop in the
example above work, albeit without concurrency for now (`translocate`
runs each child strand in arbitrary order, it could be threaded in a
future version).  In the operating system metaphor, it has no classic
equivalent, but would be bit like requesting the scheduler to run
another process specifically.