Implement Prog composition and host bootstrapping
This patch allows preparing a blank Hetzner (or other root-ssh-keyed)
server automatically.

First, make sure that you can log into the server via your `ssh-agent`
implementation.  Clover uses the agent in a non-optional capacity for
development.  Various operating systems handle this differently, but
you can usually make sure it work as intended by running:

    ssh-add

In most situations, you don't even need to do this, however; whenever
Clover in development mode tries to log into a server, it contacts
`ssh-agent` and most implementations will open an operating system
specific window to prompt for your password *once* to decrypt the
private key material and hold it in memory.

That's enough background about ssh-agent, but it's important to know
its proper function is non-optional.

Moving on, create a `Sshable` record and a `VmHost` record.  Note how
they are correlated by sharing an `id`:

    # Replace with any host name, IPv4, or IPv6 address.
    sa = Sshable.create(host: '95.216.67.38')
    vh = VmHost.create { _1.id = sa.id }

Then create a "Strand" (compare: Thread, Fiber) running the
`Vm::HostNexus` prog:

    st = Strand.create(prog: "Vm::HostNexus", label: "start") {
      _1.id = vh.id
    }

Then, you can run:

    st.run

Until the `Vm::HostNexus` ends in the `wait` label, bootstrapping the
host along the way.

== How is this organized

Clover uses metaphors from machine and operating system architecture
to organize its asynchronous processing:

* Strands are like Threads or Fibers
* Progs are like program text. They get loaded into strands.
* Parent strands, like unix processes, are responsible for reaping
  child strands
* The stack is like...the stack
* There are return values (when popping stack) and exit codes/values
  (when ending the strand)

Since this patch also introduces composition of Progs to organize
this, I will go over how it works.

First, there's some convention: there is a category of Prog, called
the "Nexus."  In this patch, it is the `Vm::HostNexus`.  The Nexus
prog is special:

* It runs on a root Strand, with no parent.  Its lifetime is roughly
  as long as the entity record itself, i.e. from creation to deletion.
* A root strand shares an id with an entity record (`VmHost`).  This
  allows you to write `VmHost.first.strand.run`, for example.
* A root strand primarily exists to start descendant progs
* A root strand waits for those progs to complete before moving on

The way those descendant progs are started is via the `bud` method.
The `bud` method accepts an initial stack frame, to be encoded as
JSONB, as a parameter.  Most of the time, initial stack contains the
id of an entity record.

N.B.: As of yet, we have no code that pushes frames onto the stack,
      save for initialization in `bud`.  The future design intention
      is that, in cases where concurrency is not desired, instead of
      creating a child strand we could push and pop frames on the
      stack instead to compose progs.  I do not envision stacks, or
      Strand parent relationships, being deeper than four elements,
      however.

The way that child strands are cleared is via `reap`.

After reaping, `leaf?` is used to determine if there are any
descendant progs left.

Within a prog, progress is recorded via `hop`, which is inspired by
the machine code idea of "jump".

`push` allows placing a prog on the stack within the concurrency
context of the strand...and when it calls pop, a return value is
propagated.

`donate` is used as a development convenience that should also be
harmless for some time in production: it allows `.run` on a `Strand`
(in practice, a root strand) to execute code in descendant strands.
The name is inspired by the concept of "priority donation" in
operating systems.

This is useful when the parent strand is waiting for child strands to
exit and has nothing else to do.  This is what makes the loop in the
example above work, albeit without concurrency for now.  `donate` runs
each child strand in arbitrary order, it could be threaded in a future
version.

== Why is it organized this way

There are some design theories to this.  None of these defects were
project-lethal in magnitude, but I aim to slightly improve with each
system I write:

1. Short progs: some previous projects suffered from some large source
   files, because they did not address both the class and the method
   to be run.  Using the `(prog, label)` attributes in Strand is one
   way to solve this.

2. Composable progs: all my previous works were effectively
   goto-based, a thicket of state transitions, i.e. used `hop` only.
   To the extent a "prog" or its analogue had two possible
   exit-transitions, I would write some state into an entity record
   and have a conditional expression.  Each case was a sui generis
   decision.

   Using a stack (although, implementation is incomplete) solves this,
   to enable structured programming in the familiar way.  Thus, the
   `(prog, label)` state that results from `pop` is dynamic and set by
   the caller, just as the instruction and stack pointer are in a C
   program.

3. Dynamic concurrency: all of my previous works made it expensive to
   make code that had no sequencing dependency concurrent.  Strand
   `bud` and `reap` (just like operating system `spawn` and `wait`)
   solves this problem.

4. Stable notation: a couple of branches in `pop` can either mark a
   strand as exited (sets `exitval`) or return to a higher stack frame
   in the same Strand via `retval`. So both composition in a
   concurrent setting and a non-concurrent setting have one notation,
   thus, there's no penalty in organizing progs into concurrent
   strands, or being used non-concurrent subroutines.

5. Performance measurement: I chose to organize the `Nexus` as a long
   lived strand spawning short-lived strands so it would be easier to
   track the time a sub-task takes to execute.  You could push the
   transient and hierarchically organized strand IDs to anything
   capable of analyzing the data, e.g. flamecharts, opentelemetry,
   bespoke scripts or SQL queries.  It's harder to do this if you have
   to find how a task "entry" state transition matches a task "exit"
   state transition, the idea of things being "beginning" and
   "finished" are less legible than when there are explicit stack or
   tree manipulation operations that we programmers emit while doing
   work in the easiest way.