Enable use of bundler and thus gems in Rhizome
It would be nice if Gemfile.lock could record content hashes of gems
to better notice a supply chain attack.

Another option is to use Bundler's vendor cache feature and commit
gems to Git, but, this is rather bloated compared to everything else
in Rhizome, so I'm not doing it for now.