Split an IPv6 network into halves for guest and Clover use
Previously, two adjacent IP addresses were used,
e.g. `2a01:4f9:2b:35b:7e40:e918::1` would be for the guest VM and
`2a01:4f9:2b:35b:7e40:e918::2` would be for Clover's internal use for,
e.g., IPsec overlay networking traffic.

I saw this as an expedient considering how much address space was
available, as it foreclosed allowing either clover or the guest VM to
freely use low bits of a larger prefix.

This patch fixes that: now one, high-order bit will be reserved for
Clover's infrastructure use.  The default I chose higher up the stack
is to delegate /95s to each network namespace hosting a guest, but
within that namespace, there is one further subdivision the lower
half, a /96 network, is sent directly to the guest, and the upper half
/96 is for Clover's use.

For Clover's purposes, the only known internal use intended is overlay
networking with IPsec, and it only needs one IP address in principle.
This has not yet been integrated, it only exists as an experiment I
did, and notes on it, after finding there was not enough mechanism in
it to abstract it.