Add Vm Host allocation state and assemble class methods
Motivated by watching engineers getting familiar with the system.

Firstly, it's hard to remember which parts of the data model to build
for a given high-level entity unless one does it all the time, so
encapsulate this in a class method, idiomatically called `assemble`.

Secondly, it's hard to remember that one needs to "prepare" a host
first, and the state of having been prepared is not that legible,
as-is you'd define it as "a VmHost that has reached the 'wait' label
in the Nexus prog at least once".  This is also hard to query in an
allocator.  So, update an enum value once a host has been prepared and
is ready to accept Vm allocations.

So, the new set-up process for a fresh VmHost is like this:

    st = Prog::Vm::HostNexus.assemble('95.216.67.38')
    st.run # loop this until st.label is "wait"

And the way to allocate a VM is:

    st = Prog::Vm::Nexus.assemble
    st.run # loop until st.label is "wait"