Do Rodauth migration with same database connection options except role
This was a problem in CI where it can't use unix auth (as in
development) and instead makes a real socket connection: passing just
dbname is not good enough in that case.

It is adequate for that use that the password on the `ph` role is the
same as the application role.  In production we are advised to make
that role NOLOGIN anyway and the preparation is quite different
anyhow.  See

https://github.com/jeremyevans/rodauth#label-Locking+Down+-28PostgreSQL+only-29