Add initial VM index and create view
By running:

    bundle exec rackup

And then visiting http://localhost:9292, you can create an account.
Check the rackup log for the verification link to navigate to, in
production, we would send that output as email.  Having verified, log
in.  You'll see the RodAuth demo pages.

To visit the VM index, visit http://localhost:9292/vm.  At first it
will be empty.

To create a vm, navigate to http://localhost:9292/vm/create

In order for a vm to allocate a slot on a host, you need a host where
`vm.location` matches `vm_host.location`.  Here are the strings I am
using in the web form:

    hetzner-hel1
    hetzner-nbg1
    equinix-da11

Because vm allocation is lazy, you can create the VmHost for a Vm
*after* the Vm record already exists.  You can create VmHost records
in a particular location like so:

    Prog::Vm::HostNexus.assemble('145.40.99.21', location: 'equinix-da11')
    Prog::Vm::HostNexus.assemble('95.216.67.38', location: 'hetzner-hel1')

It so happens that these addresses were actually in Equinix Dallas 11
and Hetzner Helsinki 1, respectively, but you are free to enter
arbitrary location data for testing.

Besides web pages and location constraints, another change in this
patch is lazy Linux image downloading.  The reason for this is I
wanted to test a second image (I chose AlmaLinux) but I did not want
to penalize every development path with downloading both images when
preparing a host, considering it is convenient to drop the Postgres
database and erase the host while developing.  In this way, the common
case where people never download AlmaLinux remains as quick as it did
before, although the time spent downloading the Linux boot image is
shifted to the first VM using that image on a host.

For demonstration purposes, it would make sense to "prewarm" the host
and download both images.