Expand VM namespace to /79 from /95
Thus, instead of a pair of /96s for each namespace, the lower
allocated to the VM and the upper allocated to Clover, we instead have
two /80s, which are five hextets in IPv6 notation rather than six.

I previously chose /96 to on the principle of being conservative and
it being easier to expand than to contract, but it still would be a
pretty big pain to do it.

But, after considering some of IPv4 transition technologies this
implementation may see in its life, one is SIIT, whereby a /96 is
delegated for the purpose of holding IPv4 addresses in its remaining
bits.

If you only have one /96 range to work with for all purposes, then you
can't reasonably delegate it wholesale, i.e. one /96 for the VM or
Clover is not big enough.  Hence, the upgrade to /80.

Several large mobile phone operators have transitioned to IPv6-only
infrastructure in this way, some almost ten years ago (T-Mobile).
Their approach is called "464XLAT." Mobile providers have a lot of
address pressure, so they are the most interested, but operating
systems and data centers may follow in the coming years.

All in all, I think SIIT techniques might be useful to both us and
customers some day, and I didn't have a great reason to be so
tight-fisted with customer address space, so, let's expand it in off
chance it can help us.