Make several improvements and fixes to SSH session caching
The most important fix is invalidating the cache when `cmd` re-raises
an error from the session.

Also, `clear_cache` didn't do the job, and it doesn't make any sense
as an instance method anyway.  Rename it to `reset_cache` since it
affects every cache line.  Add `invalidate_cache_entry` to eliminate
the single SSH session associated with the Sshabe record and call it
when re-raising.

The previous tests did not encode invariants to catch these mistakes,
so make them more powerful.