Add multi-run interface and change run's return contract
I reworked the layers here. Previously, `lease` returned whether the
lease was acquired or not, which is not useful enough.

So instead, return the `Prog::Base::FlowControl` instance that is now
obliged to be emitted by each label execution of a Strand.  This
allows the developer to see the result of each `run` better, as a
return value, and also allows chaining of several short steps together
into one lease.

Now, the developer has a concise choice.  They can run:

    st.run

To run one label as usual. But if they wish to run until the strand
pauses or exits, something like...

    st.run 30

...is conveniently at hand.

Furthermore, since the return value is more informative than before, I
removed some noisy `puts`.