Add Semaphore implementation
Semaphores are a concurrency control mechanism, in this case
implemented using database transaction rather than in memory as would
be seen in an operating system.

Generally, the way they work is this:

* Any thread of execution, e.g. web server, personnel using the REPL,
  other Strands, can increment any semaphore anywhere at any time.
  This is safe.  These could be called producers.

* Only one strand, generally, the Nexus strand, is permitted to decr.
  These could be called consumers.

A 2016 blog post about the concurrency pitfalls solved by semaphores
can be seen here for detailed information:
https://www.citusdata.com/blog/2016/08/12/state-machines-to-run-databases/

One major alteration from this blog post is that this implementation
uses a unary numbering system, e.g. previously a counter would be
incremented or decremented, but now increment makes a new row and
decrement deletes all rows in a snapshot.  Snapshots are, as a rule,
read at the beginning of a work-performing transaction to itemize the
exact amount of semaphore to decrement.

This row-heavy approach bloats the size of each semaphore increment
hugely (from less than a bit each to many bytes each) but offers
potential for progress tracking (a long-standing complaint with a
counter-only approach) and finer-grained locking in the context of a
relational database.

We can hybridize approaches later by adding a counter to the records
and incrementing it in some situations, if we need increments and
decrements to be cheaper in the storage dimension.  The downside is
that a long transaction that has incremented the counter but not yet
committed squats on the lock, and more software engineering is
required to defer modifications of semaphores so the locks are held
briefly.