Upgrade to cloud-hypervisor to v31.0 from v30.0
While doing so, I out-lined the cloud-hypervisor version to a
constant, as I had to do a couple of integration tests to detect
issues in what was otherwise a simple change.

I was updating cloud-hypervisor and endured an unnecessary mismatch in
the part of the program that runs it vs. the part that downloads it.