Improve erb-formatter integration into the development process
Add it to CI and switch to a branch that does not use multi-line class
attributes, to be less verbose when working with "utility" CSS
frameworks.

A slight change is made to rubocop rules not out of style preference,
but because erb-formatter and rubocop disagree on code formatting on
at least this one minor point.  This patch opts to let erb-formatter
have it for simplicity reasons.