Switch to edk2 firmware
This firmware seems more fully featured.  It has some promise of
supporting Secure Boot, though I have not enabled that in this build
yet.  Also it functions with the OpenSUSE Leap 15.4 cloud qcow2 image,
like, at all.  And Aarch64.

I compiled this on my laptop.  It's not the most Pentagon-esque setup
we could have for producing binary artifacts, but it's a start.