Learn VmHost core and memory quantities during prep
This allows us to compute a memory to core ratio and then divvy up
resources in proportion to that.  It also adds a field to record how
many cores have already been allocated.

Here I experiment with using `exitval` for the first time.  Note how
this decouples LearnMemory and LearnCores from VmHost: the data is
passed out of the prog via the exitval, and could be used by any
purpose, so long as there is an Sshable record (e.g. if there was a Vm
that we sought to administrate via Sshable).

We have no need for reuse of those at present, so normally I'd frown
on what could be considered an aesthetic choice, but I'm still in the
exploration phase on some of the core features of the Strand execution
system and do not quite yet know what I'm doing with some of the
constructs.  This is part of getting there.

The two new progs have 100% branch coverage, and achieving this was
instructive.  It was mostly fine, though.

Alas, I have to relax Model freezing in test, to allow `expect` to
function for something like `expect(Sshable).to receive(:[])`.