Implement core-wise VM sizing
This scales the number of CPU threads, usage of NUMA nodes, and memory
to the core level being purchased.  There are also improvements
(e.g. avoiding over-allocation) and simplifications (via the
materialized `used_cores` field) to the VmHost allocator, which now
also boasts many more tests.

I had to disable a RuboCop recommendation, as otherwise there's no way
for a method to remember anything for auto-increment to get unique
values.  I do things this way because randomness proves surprisingly
expensive if used more often than necessary in (ever-growing) tests.

As expedient, read the number of cores out of the "size," which is
also used for UI rendering: a "standard-1" is one core, a
"standard-20" is 20 cores.

There are unfortunately has a lot of nomenclature disagreements.  In
most situations, because of its popularity, I use Linux nomenclature,
but when getting closer to dealing with Cloud Hypervisor, I use its
nomenclature and centralize the conversion in terms.

Here's how they match up respectively:

* Linux: cpus, cores, nodes, sockets
* Cloud Hypervisor: vcpu or threads, cores, dies, packages

The general idea here is to allocate VMs core by core, the smallest
unit that makes sense.  Allocating thread-by-thread does not, since
the general idea of those threads is to squeeze a bit more pipeline
efficiency by increasing the concurrency level exposed to the
operating system; as such, the microarchitectural contention on the
core they share is very apparent.

There are two major considerations for this:

* Getting the correct number of vcpus passed, accounting for core
  "threads".  For most server configurations on x64, this is a 2:1
  ratio.

* Passing the correct NUMA (nodes a.k.a. dies) topology to the VM, and
  trying to avoid exposing a difficult NUMA topology.

The subject of NUMA will get very hairy, especially once we start
backing VMs with hugepages and those hugepages are resident to
particular NUMA nodes.  But seeing as I was groveling through `lscpu`
output and writing formula to fill in the thread topology information
in Cloud Hypervisor -- it's all or nothing -- I decided to get basic
handling of NUMA in place, rather than having to make a remark that
this is an invisible lurking consideration every time we discuss
constraints on the allocation system.