Allow zero-length naps to execute again immediately
Although "sleep 0" has traditionally been a way for unix programs to
advise a good time to yield execution, breaking the "run" loop does
make `st.run 45` annoying with zero-length naps which are emitted from
"donate" in multi-strand progs.

So go ahead and keep running instead.

I considered making this dependent on whether the session was a REPL
or not, but I don't see the obvious harm of having this behavior in
`bin/respirate` as well.