Fix fragile tests for zero-length naps
This should have been a part of
f5b67140e44af5305248b8f49dbbb3c7127e0194, otherwise, the tests fail in
non-deterministic fashion (as CI did succeed).

The tests were not designed for dispatcher to spin running the label
in question multiple times because "nap 0" has become a loop,
effectively, rather than interrupting execution as it was before,
which was the point, to make running strands with "donate" which emits
a zero-length nap more convenient.