Compile and install dnsmasq on host prep
As dnsmasq may offer a passable alternative to setting up separate
DHCP, DNS (for some cases...), and IPv6 Router Advertisement
infrastructure, I decided to integrate it into the stack.

As it turns out, I think we may want very new and possibly modified
versions, so I wrote this prog to clone it from a copy of mine on
GitHub, compile, and install it.

This prog does something new not seen in the code base: it buds
*itself* to perform parallel steps, and multiple labels exit via
`pop`.

This is an interesting pattern: the `Vm::HostNexus` is makes extensive
use of small, supervised progs to maintain focus in the nexus file,
but introduces some bloat in the process.  The `Vm::Nexus` takes the
opposite extreme, doing no concurrency and being linear, using `hop`
more extensively to chain things together, more akin to Clover's
predecessors.

This does something new, with most of the compactness of `hop` but the
concurrency of `bud.` And atop of that, it slots in neatly and
concurrently with `Vm::HostNexus`, whose focus on smaller files and
concurrency is looking better and better as it grows.  In doing so,
it's our first Strand supervision chain that is thrice deep:
`Vm::HostNexus`, supervising `InstallDnsmasq`, which in turn
supervises copies of itself that start from a different label (in
principle initial stacks could also be used to vary the entry point).