Call InstallRhizome through BootstrapRhizome
This shortens Vm::HostNexus and makes it easier to re-use the two-step
of bootstrapping a Sshable with the rhizome user and installing the
source all at once, as would be the case in a Vm that clover
administrates with SSH (e.g., a MinIO cluster member).

It may seem odd to install all of Rhizome (which includes many
Vm/hypervisor parts) in situations where we do no such work, but the
number of bytes is fairly small and some of the library needs will be
held in common.  I think we can exercise an option to send more
fine-grained programs at some later time when we understand what is
shared and how to structure that.