Lease out exited strands to delete their semaphores
If an exited strand has some other thread increment a semaphore, it
will never delete that semaphore, and will get stuck on foreign keys
while trying to delete.  So it is necessary to lease an exited strand
if only to wipe out semaphores.

A fix to the fix in 2c36a0add37cd2a7733d42a9dffff4c425a4ee74