Improve and simplify navigation convenience infrastructure
This patch manages to simplify the mechanism in
3ed62e21f43367a88982f47db0ba60e0570ed009: it is less dynamic, with
class definition rather than `instance_eval`.

The main thing lost here is accessors for any record, including ones
not directly correlated by `strand.id` or the stack frame's
`subject_id`.  But, in spite of that loss, I think this patch is an
improvement, for these reasons:

1) Navigation of indirectly correlated records is nonexistent in the
   code base so far. I think there are solutions, but let's wait for
   concrete ugly examples to rectify.

2) The stack-scanning mechanism encourages duplication of references
   in both entity records, and the stack.  The latter does not have
   the benefit of constraints, either.  This is asking for trouble
   when entities and a stack do not agree, and makes operational fixes
   more complex, because two references must be updated.

3) Source code has fewer subexpressions, as `bud` and `push` calls are
   simplified in all extant situations.