Fill in SSH features
Add SSH key generation and rotation, though the latter is not
automatically scheduled yet.

Key generation simplifies creating Sshable VMs in an environment that
does not have an ssh-agent with the keys loaded, i.e. non-development.

The approach here is to store tiny, bare ed25519 private key material
in encrypted columns, and bloat it up to the format expected by
OpenSSH and net-ssh on-demand.

Also, consult the ssh agent via net-ssh in the `vm_host` route, and
guard it so it only runs this convenience development mode.