Make Heroku hosting related changes
Add a Procfile, the most Heroku-idiosyncratic thing.  Though, it can
also be useful: with foreman, you can run:

   foreman start -c web=1,respirate=1

To run a web and respirate process.

Pry is made a production dependency, since the REPL is useful to have
with `heroku run`.

I also add Tailwind compilation to assets:precompile, which is run by
the Heroku Ruby buildpack, tighten up some cookies except in test and
development environments, and add a dependency on puma, since we had
no non-development server previously.