Use Unreloader a bit more mindfully
It doesn't make sense to only run it in development, otherwise routes
like `vm/create` are not installed in non-development
environments...including test.

Note that routes are already ignored by Zeitwerk in loader.rb.

There's also a load of a non-existent file that, for reasons I do not
fully understand, did not cause problems previously.

Still, we have a long way to go in figuring out how to compose
Zeitwerk and Unreloader or how to use only one or the other.  It is
awfully nice that Unreloader lets you change files and refresh a page
for checking web work.