Switch from static network configuration to dnsmasq and dhcp6
Of the three distributions supported, only Ubuntu has been successful
in obtaining intended network access on the first boot.  Now AlmaLinux
is also successful, but openSUSE still doesn't work.

The main errata here is this breaks support for multiple IPv6 prefixes
on an interface for now.  This is required for the mesh networking
support.  Someone, probably me, will add those back in later.

A side effect is also now, much like Hetzner, the third IPv6 address
in the subnet, ::2, is allocated to the VM.  dnsmasq appears to
require an in-subnet address on the host side to listen on for DHCP
messages, so I chose the customary "gateway" address at ::1. Later on,
it can also be the local DNS server, since we want to offer in-mesh
DNS names not available in the wider world as well.  My attempt to ask
a question about avoiding this intrusion into the customer's subnet
was met with silence so far:

https://lists.thekelleys.org.uk/pipermail/dnsmasq-discuss/2023q2/017013.html

The zero-suffix address, `::`, is left empty: this is also apparently
an IPv6 norm to avoid confusion between referencing a network and
referencing an address & prefix length combo.  In principle, however,
the guest VM can bind it and use it if they want, though.

Background information:

There were some reports on the web about static networking cloud-init
not quite working, perhaps only Ubuntu (whose sponsoring company also
controls cloud-init) took care to make it work at this time.

In any case, static network configuration was never going to remain
good enough for very long: for one, we can't boot a VM on another
machine, since cloud-init nominally runs just once...to init.